[LO-Inheritance] Attempted to importa common methods
